Move more internal types to the im module
