+++ rs-matter/src/dm/types/reply.rs
@@ -15,20 +15,11 @@
  *    limitations under the License.
  */
 
-use crate::acl::Accessor;
-use crate::dm::{AsyncHandler, HandlerContext, Node};
-use crate::error::{Error, ErrorCode};
-use crate::im::encoding::{
-    AttrDataTag, AttrPath, AttrResp, AttrRespTag, AttrStatus, CmdDataTag, CmdPath, CmdResp,
-    CmdRespTag, CmdStatus, EventData, EventFilter, EventPath, EventResp, IMStatusCode,
-};
-use crate::tlv::{TLVArray, TLVElement, TLVTag, TLVWrite, TagType, ToTLV};
-use crate::transport::exchange::Exchange;
-use crate::utils::storage::WriteBuf;
+use crate::error::Error;
+use crate::im::encoding::{AttrDataTag, AttrPath, AttrRespTag, CmdDataTag, CmdPath, CmdRespTag};
+use crate::tlv::{TLVTag, TLVWrite, TagType, ToTLV};
 
-use super::{
-    AttrDetails, CmdDetails, InvokeContextInstance, ReadContextInstance, WriteContextInstance,
-};
+use super::{AttrDetails, CmdDetails};
 
 // A type for writing the outcome of an attribute-read or command-invoke operation.
 pub trait Reply {
@@ -66,424 +57,6 @@ pub trait ReadReply {
 pub trait InvokeReply {
     /// Return the reply data writer for a command with the provided Command ID.
     fn with_command(self, cmd: u32) -> Result<impl Reply, Error>;
-}
-
-pub struct HandlerInvoker<'a, 'b, C> {
-    exchange: &'b mut Exchange<'a>,
-    context: C,
-}
-
-impl<'a, 'b, C> HandlerInvoker<'a, 'b, C>
-where
-    C: HandlerContext,
-{
-    pub const fn new(exchange: &'b mut Exchange<'a>, context: C) -> Self {
-        Self { exchange, context }
-    }
-
-    pub fn exchange(&mut self) -> &mut Exchange<'a> {
-        self.exchange
-    }
-
-    pub async fn process_read(
-        &mut self,
-        item: &Result<AttrDetails, AttrStatus>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        let tail = tw.get_tail();
-
-        let result = self.do_process_read(item, &mut *tw).await;
-
-        if result.is_err() {
-            // If there was an error, rewind to the tail so we don't write any data.
-            tw.rewind_to(tail);
-        }
-
-        result
-    }
-
-    async fn do_process_read(
-        &mut self,
-        item: &Result<AttrDetails, AttrStatus>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        let result = match item {
-            Ok(attr) => {
-                let pos = tw.get_tail();
-
-                let result = self.read(attr, &mut *tw).await;
-
-                match result {
-                    Ok(()) => Ok(None),
-                    Err(e) if e.code() != ErrorCode::NoSpace => {
-                        error!("Error reading attribute: {}", e);
-
-                        tw.rewind_to(pos);
-
-                        Ok(attr.status(e.into()))
-                    }
-                    Err(e) => Err(e),
-                }
-            }
-            Err(status) => {
-                error!("Error processing attribute read: {:?}", status);
-                Ok(Some(status.clone()))
-            }
-        };
-
-        match result {
-            Ok(Some(status)) => AttrResp::Status(status).to_tlv(&TagType::Anonymous, tw),
-            Ok(None) => Ok(()),
-            Err(err) => Err(err),
-        }
-    }
-
-    pub async fn read(&mut self, attr: &AttrDetails, tw: &mut WriteBuf<'_>) -> Result<(), Error> {
-        self.context
-            .handler()
-            .read(
-                ReadContextInstance::new(self.exchange, &self.context, attr),
-                ReadReplyInstance::new(attr, tw),
-            )
-            .await
-    }
-
-    pub async fn process_write(
-        &mut self,
-        item: &Result<(AttrDetails, TLVElement<'_>), AttrStatus>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        let tail = tw.get_tail();
-
-        let result = self.do_process_write(item, &mut *tw).await;
-
-        if result.is_err() {
-            // If there was an error, rewind to the tail so we don't write any data.
-            tw.rewind_to(tail);
-        }
-
-        result
-    }
-
-    async fn do_process_write(
-        &mut self,
-        item: &Result<(AttrDetails, TLVElement<'_>), AttrStatus>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        let result = match item {
-            Ok((attr, data)) => {
-                let pos = tw.get_tail();
-
-                let result = self.write(attr, data).await;
-
-                match result {
-                    Ok(()) => {
-                        // A write that was accepted by the cluster handler
-                        // counts as an attribute change for subscription
-                        // reporting purposes. Notify generically here so that
-                        // cluster handlers do not each need to call
-                        // `notify_attr_changed` from every attribute setter.
-                        self.context.notify_attr_changed(
-                            attr.endpoint_id,
-                            attr.cluster_id,
-                            attr.attr_id,
-                        );
-                        Ok(attr.status(IMStatusCode::Success))
-                    }
-                    Err(err) if err.code() != ErrorCode::NoSpace => {
-                        error!("Error writing attribute: {}", err);
-
-                        tw.rewind_to(pos);
-
-                        Ok(attr.status(err.into()))
-                    }
-                    Err(err) => Err(err),
-                }
-            }
-            Err(status) => {
-                error!("Error processing attribute write: {:?}", status);
-                Ok(Some(status.clone()))
-            }
-        };
-
-        match result {
-            Ok(Some(status)) => status.to_tlv(&TagType::Anonymous, tw),
-            Ok(None) => Ok(()),
-            Err(err) => Err(err),
-        }
-    }
-
-    pub async fn write(&mut self, attr: &AttrDetails, data: &TLVElement<'_>) -> Result<(), Error> {
-        self.context
-            .handler()
-            .write(WriteContextInstance::new(
-                self.exchange,
-                &self.context,
-                attr,
-                data,
-            ))
-            .await
-    }
-
-    pub async fn process_invoke(
-        &mut self,
-        item: &Result<(CmdDetails, TLVElement<'_>), CmdStatus>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        let tail = tw.get_tail();
-
-        let result = self.do_process_invoke(item, &mut *tw).await;
-
-        if result.is_err() {
-            // If there was an error, rewind to the tail so we don't write any data.
-            tw.rewind_to(tail);
-        }
-
-        result
-    }
-
-    async fn do_process_invoke(
-        &mut self,
-        item: &Result<(CmdDetails, TLVElement<'_>), CmdStatus>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        let result = match item {
-            Ok((cmd, data)) => {
-                let pos = tw.get_tail();
-
-                let result = self.invoke(cmd, data, &mut *tw).await;
-
-                match result {
-                    Ok(()) => {
-                        if pos == tw.get_tail() {
-                            Ok(cmd.status(IMStatusCode::Success))
-                        } else {
-                            Ok(None)
-                        }
-                    }
-                    Err(err) if err.code() != ErrorCode::NoSpace => {
-                        error!("Error invoking command: {}", err);
-
-                        tw.rewind_to(pos);
-
-                        Ok(cmd.status(err.into()))
-                    }
-                    Err(err) => Err(err),
-                }
-            }
-            Err(status) => {
-                error!("Error processing command: {:?}", status);
-                Ok(Some(status.clone()))
-            }
-        };
-
-        match result {
-            Ok(Some(status)) => CmdResp::Status(status).to_tlv(&TagType::Anonymous, tw),
-            Ok(None) => Ok(()),
-            Err(err) => Err(err),
-        }
-    }
-
-    pub async fn invoke(
-        &mut self,
-        cmd: &CmdDetails,
-        data: &TLVElement<'_>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<(), Error> {
-        self.context
-            .handler()
-            .invoke(
-                InvokeContextInstance::new(self.exchange, &self.context, cmd, data),
-                InvokeReplyInstance::new(cmd, tw),
-            )
-            .await
-    }
-}
-
-pub struct EventReader {
-    max_seen_event_number: u64,
-    next_max_seen_event_number: u64,
-    /// Whether the originating Read/Subscribe request had `fabricFiltered=true`.
-    /// When set, fabric-sensitive events (those whose payload carries a
-    /// `FabricIndex` context-tag 254) are dropped if their fabric index does
-    /// not match the accessor's. See Matter Core spec.
-    fabric_filtered: bool,
-}
-
-impl EventReader {
-    pub const fn new(
-        max_seen_event_number: u64,
-        next_max_seen_event_number: u64,
-        fabric_filtered: bool,
-    ) -> Self {
-        Self {
-            max_seen_event_number,
-            next_max_seen_event_number,
-            fabric_filtered,
-        }
-    }
-
-    pub fn process_read(
-        &mut self,
-        event: EventData<'_>,
-        paths: &TLVArray<'_, EventPath>,
-        event_filters: &Option<TLVArray<'_, EventFilter>>,
-        node: &Node<'_>,
-        accessor: &Accessor<'_>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<bool, Error> {
-        let event_number = event.event_number;
-        if !(event_number > self.max_seen_event_number
-            && event_number <= self.next_max_seen_event_number)
-        {
-            // This event is outside the range of interest, skip
-            return Ok(false);
-        }
-
-        let tail = tw.get_tail();
-
-        let result = self.do_process_read(event, paths, event_filters, node, accessor, &mut *tw);
-
-        if result.is_err() {
-            // If there was an error, rewind to the tail so we don't write any data
-            // and leave `max_seen_event_number` untouched so this event will be
-            // retried on the next chunk.
-            tw.rewind_to(tail);
-        } else {
-            // The event was considered (whether or not it actually matched the
-            // path/filter/access checks). Advance the local watermark so that
-            // chunked reads do not re-consider the same event again on
-            // continuation, and so that the iteration converges.
-            self.max_seen_event_number = event_number;
-        }
-
-        result
-    }
-
-    fn do_process_read(
-        &mut self,
-        event: EventData<'_>,
-        paths: &TLVArray<'_, EventPath>,
-        event_filters: &Option<TLVArray<'_, EventFilter>>,
-        node: &Node<'_>,
-        accessor: &Accessor<'_>,
-        tw: &mut WriteBuf<'_>,
-    ) -> Result<bool, Error> {
-        if self.fabric_filtered && !Self::matches_fabric(&event, accessor) {
-            return Ok(false);
-        }
-
-        if Self::matches_paths(&event, paths, node, accessor)?
-            && Self::matches_filters(&event, event_filters)?
-            && Self::matches_access(&event, node, accessor)?
-        {
-            EventResp::Data(event).to_tlv(&TagType::Anonymous, &mut *tw)?;
-
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Per Matter Core spec (Fabric-Sensitive Reporting):
-    /// When `fabricFiltered=true`, fabric-sensitive events (those whose payload
-    /// carries a `FabricIndex` field at context tag 254) SHALL only be reported
-    /// to the requesting fabric.
-    ///
-    /// Events without a `FabricIndex` field are not fabric-sensitive and pass
-    /// through unfiltered. Events with a `FabricIndex` field that matches the
-    /// accessor's fabric are reported as well.
-    fn matches_fabric(event: &EventData<'_>, accessor: &Accessor<'_>) -> bool {
-        // Inspect the event payload struct for context tag 254 (FabricIndex).
-        let Ok(payload) = event.data.structure() else {
-            // Not a struct payload — treat as non-fabric-sensitive.
-            return true;
-        };
-
-        let Ok(elem) = payload.find_ctx(crate::im::encoding::FABRIC_INDEX_TAG) else {
-            // No `FabricIndex` field — non-fabric-sensitive, allow.
-            return true;
-        };
-
-        match elem.non_empty().and_then(|e| e.u8().ok()) {
-            Some(fab_idx) => fab_idx == accessor.fab_idx,
-            // Field present but unreadable / null — be conservative and allow.
-            None => true,
-        }
-    }
-
-    fn matches_paths(
-        event: &EventData<'_>,
-        paths: &TLVArray<'_, EventPath>,
-        node: &Node<'_>,
-        accessor: &Accessor<'_>,
-    ) -> Result<bool, Error> {
-        for path in paths {
-            let path = path?;
-
-            if Self::matches_path(event, path, node, accessor) {
-                return Ok(true);
-            }
-        }
-
-        Ok(false)
-    }
-
-    fn matches_path(
-        event: &EventData<'_>,
-        path: EventPath,
-        node: &Node<'_>,
-        accessor: &Accessor<'_>,
-    ) -> bool {
-        if node.validate_event_path(&path, accessor).is_err() {
-            return false;
-        }
-
-        let epath = &event.path;
-
-        epath
-            .node
-            .is_none_or(|node| path.node.is_none_or(|expected_node| expected_node == node))
-            && epath.endpoint.is_none_or(|endpoint| {
-                path.endpoint
-                    .is_none_or(|expected_endpoint| expected_endpoint == endpoint)
-            })
-            && epath.cluster.is_none_or(|cluster| {
-                path.cluster
-                    .is_none_or(|expected_cluster| expected_cluster == cluster)
-            })
-            && epath.event.is_none_or(|event| {
-                path.event
-                    .is_none_or(|expected_event| expected_event == event)
-            })
-    }
-
-    fn matches_filters(
-        event: &EventData<'_>,
-        event_filters: &Option<TLVArray<'_, EventFilter>>,
-    ) -> Result<bool, Error> {
-        if let Some(filters) = &event_filters {
-            // Check if the event passes the filters. If it doesn't pass any of them, skip it.
-            // We assume the 99% case is that there is a single filter, on event-no, so just brute force filtering
-            for filter in filters {
-                if let Some(event_min) = filter?.event_min {
-                    if event.event_number < event_min {
-                        return Ok(false);
-                    }
-                }
-            }
-        }
-
-        Ok(true)
-    }
-
-    fn matches_access(
-        event: &EventData<'_>,
-        node: &Node<'_>,
-        accessor: &Accessor<'_>,
-    ) -> Result<bool, Error> {
-        Ok(node.validate_event_path(&event.path, accessor).is_ok())
-    }
 }
 
 /// A concrete implementation of the `ReadReply` trait for encoding attribute data.
